rename and remove unused code
